requirements.txt
# Function dependencies, for example:
# package>=version

firebase-admin
line-bot-sdk
openai
